Manually unboxing is unnecessary.  It is done automatically.

+++ NewsPetJava/src/edu/iastate/coms/cs472/newspet/reader/ItemClassificationJob.java
@@ -120,7 +120,7 @@ public class ItemClassificationJob implements Runnable
 		Pair<Integer, Double> best = null;
 		for(Pair<Integer, Double> candidate : categoryProbabilities)
 		{
-			if(best == null || candidate.getB().doubleValue() > best.getB().doubleValue()) best = candidate;
+			if(best == null || candidate.getB() > best.getB()) best = candidate;
 		}
 		return best;
 	}
